move network codes to 'net' dir
